Keychip: add A700X short data sheet
The full data sheet is under a NDA. I could not find it on the Internet.

There are more recent versions of this preliminary short data sheet, but the one
linked here is the one with the most information. Version 3.1 (which appears to
be the most recent) removed the chapter about pinning information.